Rename files and classes JC-645
