+++ apps/angular/src/app/features/anime-catalog/anime-catalog.component.html
