+++ apps/angular/src/app/features/anime-table/anime-table.component.html
