+++ apps/angular/src/assets/anime-icon.png
