+++ libs/core/dtos/anime-list.dto.ts
