+++ libs/core/dtos/error-response.dto.ts
