+++ libs/core/dtos/get-anime.dto.ts
@@ -1,0 +1,4 @@
+import { AnimeListDto } from './anime-list.dto';
+import { ErrorResponseDto } from './error-response.dto';
+
+export type GetAnimeDto = AnimeListDto | ErrorResponseDto;
