--- libs/core/dtos/get_anime.dto.ts
@@ -1,4 +1,0 @@
-import { AnimeListDto } from './anime_list.dto';
-import { ErrorResponseDto } from './error_response.dto';
-
-export type GetAnimeDto = AnimeListDto | ErrorResponseDto;
