--- libs/core/dtos/get_anime_by_id.dto.ts
@@ -1,4 +1,0 @@
-import { AnimeDetailedDto } from './anime_detailed.dto';
-import { ErrorResponseDto } from './error_response.dto';
-
-export type getAnimeByIdDto = AnimeDetailedDto | ErrorResponseDto;
