--- libs/core/dtos/studio.dto.ts
@@ -1,18 +1,0 @@
-/** Total count of items. */
-export type StudioDto = {
-
-	/** Total count of items. */
-	readonly id: number;
-
-	/** Total count of items. */
-	readonly created: string;
-
-	/** Total count of items. */
-	readonly modified: string;
-
-	/** Total count of items. */
-	readonly name: string;
-
-	/** Total count of items. */
-	readonly image: string;
-};
